Refactor: global 패키지 제외한 모든 도메인을 domain 패키지 하위로 이동
