--- src/main/java/org/sopt/article/domain/SearchType.java
@@ -1,6 +1,0 @@
-package org.sopt.article.domain;
-
-public enum SearchType {
-    TITLE,
-    MEMBER
-}
