--- src/main/java/org/sopt/article/domain/Tag.java
@@ -1,8 +1,0 @@
-package org.sopt.article.domain;
-
-public enum Tag {
-    CS,
-    DB,
-    SPRING,
-    ETC
-}
