--- src/main/java/org/sopt/article/repository/ArticleRepositoryCustom.java
@@ -1,9 +1,0 @@
-package org.sopt.article.repository;
-
-import org.sopt.article.domain.Article;
-import org.sopt.article.domain.SearchType;
-import java.util.List;
-
-public interface ArticleRepositoryCustom {
-    List<Article> search(SearchType type, String keyword);
-}
