--- src/main/java/org/sopt/article/service/ArticleService.java
@@ -1,14 +1,0 @@
-package org.sopt.article.service;
-
-import org.sopt.article.domain.Article;
-import org.sopt.article.domain.SearchType;
-import org.sopt.article.dto.ArticleCreateRequest;
-
-import java.util.List;
-
-public interface ArticleService {
-    Article create(Long memberId, ArticleCreateRequest request);
-    Article findById(Long articleId);
-    List<Article> findAll();
-    List<Article> search(SearchType title, String memberName);
-}
