+++ src/main/java/org/sopt/domain/article/domain/SearchType.java
@@ -1,0 +1,6 @@
+package org.sopt.domain.article.domain;
+
+public enum SearchType {
+    TITLE,
+    MEMBER
+}
