+++ src/main/java/org/sopt/domain/article/domain/Tag.java
@@ -1,0 +1,8 @@
+package org.sopt.domain.article.domain;
+
+public enum Tag {
+    CS,
+    DB,
+    SPRING,
+    ETC
+}
