+++ src/main/java/org/sopt/domain/article/exception/DuplicateArticleTitleException.java
@@ -1,0 +1,14 @@
+package org.sopt.domain.article.exception;
+
+import org.sopt.global.code.ErrorCode;
+import org.sopt.global.exception.BusinessException;
+
+public class DuplicateArticleTitleException extends BusinessException {
+    public DuplicateArticleTitleException(String message) {
+        super(ErrorCode.DUPLICATE_ARTICLE_TITLE);
+    }
+
+    public DuplicateArticleTitleException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
