+++ src/main/java/org/sopt/domain/article/service/ArticleService.java
@@ -1,0 +1,14 @@
+package org.sopt.domain.article.service;
+
+import org.sopt.domain.article.domain.Article;
+import org.sopt.domain.article.domain.SearchType;
+import org.sopt.domain.article.dto.ArticleCreateRequest;
+
+import java.util.List;
+
+public interface ArticleService {
+    Article create(Long memberId, ArticleCreateRequest request);
+    Article findById(Long articleId);
+    List<Article> findAll();
+    List<Article> search(SearchType title, String memberName);
+}
