+++ src/main/java/org/sopt/domain/member/domain/Gender.java
@@ -1,0 +1,7 @@
+package org.sopt.domain.member.domain;
+
+public enum Gender {
+    MALE,
+    FEMALE,
+    OTHER
+}
