+++ src/main/java/org/sopt/domain/member/domain/Provider.java
@@ -1,0 +1,8 @@
+package org.sopt.domain.member.domain;
+
+public enum Provider {
+    LOCAL,      // 일반 회원가입
+    GOOGLE,     // 구글 소셜 로그인
+    KAKAO,      // 카카오 소셜 로그인 (향후 추가)
+    NAVER       // 네이버 소셜 로그인 (향후 추가)
+}
