+++ src/main/java/org/sopt/domain/member/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package org.sopt.domain.member.repository;
+
+import org.sopt.domain.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByEmail(String email);
+}
