--- src/main/java/org/sopt/member/domain/Gender.java
@@ -1,7 +1,0 @@
-package org.sopt.member.domain;
-
-public enum Gender {
-    MALE,
-    FEMALE,
-    OTHER
-}
