--- src/main/java/org/sopt/member/exception/DuplicateMemberException.java
@@ -1,14 +1,0 @@
-package org.sopt.member.exception;
-
-import org.sopt.global.code.ErrorCode;
-import org.sopt.global.exception.BusinessException;
-
-public class DuplicateMemberException extends BusinessException {
-    public DuplicateMemberException(String message) {
-        super(ErrorCode.DUPLICATE_MEMBER);
-    }
-
-    public DuplicateMemberException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
