--- src/main/java/org/sopt/member/exception/MemberAgeException.java
@@ -1,14 +1,0 @@
-package org.sopt.member.exception;
-
-import org.sopt.global.code.ErrorCode;
-import org.sopt.global.exception.BusinessException;
-
-public class MemberAgeException extends BusinessException {
-    public MemberAgeException(String message) {
-        super(ErrorCode.MEMBER_AGE_INVALID);
-    }
-
-    public MemberAgeException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
