--- src/main/java/org/sopt/member/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package org.sopt.member.repository;
-
-import org.sopt.member.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
-}
