--- src/main/java/org/sopt/member/service/MemberService.java
@@ -1,13 +1,0 @@
-package org.sopt.member.service;
-
-import org.sopt.member.domain.Member;
-import org.sopt.member.dto.MemberCreateRequest;
-
-import java.util.List;
-
-public interface MemberService {
-    Member join(MemberCreateRequest request);
-    Member findOne(Long memberId);
-    List<Member> findAllMembers();
-    void deleteMember(Long memberId);
-}
